requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
uvicorn==0.33.0
flask==3.0.3
Flask-Cors==5.0.0
gevent==24.2.1
pymongo==4.10.1
pandas==2.0.3
chardet==5.2.0
python-dotenv==1.0.1
psycopg2-binary==2.8.6